finished things up

+++ frontend/src/components/ListA.js
@@ -9,6 +9,8 @@ import UpdateA from "./UpdateA";
 
 function ListA() {
     const [announcements, setAnnouncements] = useState([]);
+    const user = JSON.parse(localStorage.getItem("userinfo"));
+
     //const { id } = useParams();
     //console.log(id);
     const getAnnouncements = async() => {
@@ -41,24 +43,62 @@ function ListA() {
     }, []); //ensure we only make one request
     // console.log(posts)
 
-    return (
-        <div className="container">
-            <div className="row">
-            {/* <span className="border border-2"></span> */}
-                <div className="col">
-                    {announcements.map(Announcement =>  
-                        <>
-                            <hr></hr>
-                            <h1 key={Announcement.Announcement_ID}>{Announcement.Announcement_Text} </h1>
-                            <UpdateA Announcement={Announcement} />
-                            <button onClick={() => deleteAnnouncement(Announcement.Announcement_ID)}>Delete</button> 
-                        </>
-                    )}
+    if ("userinfo" in localStorage) {
+        if (user.isAdmin) {
+            return (
+                <div className="container">
+                    <div className="row">
+                        {/* <span className="border border-2"></span> */}
+                        <div className="col">
+                            {announcements.map(Announcement =>
+                                <>
+                                    <hr></hr>
+                                    <h1 key={Announcement.Announcement_ID}>{Announcement.Announcement_Text} </h1>
+                                    <UpdateA Announcement={Announcement} />
+                                    <button onClick={() => deleteAnnouncement(Announcement.Announcement_ID)}>Delete</button>
+                                </>
+                            )}
+                        </div>
+                    </div>
+                </div>
+
+            );
+        } else {
+            return (
+                <div className="container">
+                    <div className="row">
+                        {/* <span className="border border-2"></span> */}
+                        <div className="col">
+                            {announcements.map(Announcement =>
+                                <>
+                                    <hr></hr>
+                                    <h1 key={Announcement.Announcement_ID}>{Announcement.Announcement_Text} </h1>
+                                </>
+                            )}
+                        </div>
+                    </div>
+                </div>
+
+            );
+        }
+    } else {
+        return (
+            <div className="container">
+                <div className="row">
+                    {/* <span className="border border-2"></span> */}
+                    <div className="col">
+                        {announcements.map(Announcement =>
+                            <>
+                                <hr></hr>
+                                <h1 key={Announcement.Announcement_ID}>{Announcement.Announcement_Text} </h1>
+                            </>
+                        )}
+                    </div>
                 </div>
             </div>
-        </div>
 
-    );
+        );
+    }
 }
